requirements.txt
torch>=2.0.0,<3
transformers[torch]>=4.30.0,<5
pillow>=11.1.0
httpx==0.27.2
openai==1.79.0
pillow-avif-plugin==1.5.2
pi-heif==1.0.0
google-cloud-storage